# swagger.yaml
openapi: 3.0.3
info:
  title: Girls Got Game API
  description: Basketball training application API for mobile and web clients
  version: 2.0.0
  contact:
    name: Girls Got Game API Support
    email: support@girlsgotgame.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.girlsgotgame.app/api
    description: Production server
security:
  - cookieAuth: []
paths:
...
  /mobile-auth/sign-in/mobile:
    post:
      tags:
        - Mobile Authentication
      summary: Mobile OAuth sign-in
      description: Authenticate mobile users with Google ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [google]
                  example: google
                idToken:
                  type: string
                  description: Google ID token from mobile OAuth
              required:
                - provider
                - idToken
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  session:
                    $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /mobile-auth/sign-out/mobile:
    post:
      tags:
        - Mobile Authentication
      summary: Mobile sign out
      description: Sign out mobile user and invalidate session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Successfully signed out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Profile Endpoints
...
  /profiles:
    post:
      tags:
        - Profiles
      summary: Create or update profile
      description: Create or update user profile during onboarding
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfile'
      responses:
        '201':
          description: Profile created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /workouts:
    get:
      tags:
        - Workouts
      summary: Get user workouts
      description: Retrieve workouts for the authenticated user
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Workouts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Workouts
      summary: Create workout
      description: Create a new workout entry
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkout'
      responses:
        '201':
          description: Workout created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /posts:
    post:
      tags:
        - Posts
      summary: Create post
      description: Create a new social media post
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostWithRelations'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /upload/media:
    post:
      tags:
        - Upload
      summary: Upload media
      description: Upload media for feed posts (128MB limit)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Media uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUpload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Chat Endpoints
...
  /games:
    get:
      tags:
        - Games
      summary: Get all games
      description: Retrieve all games (public endpoint)
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Games
      summary: Create game
      description: Create a new game (admin only)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGame'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /media/upload:
    post:
      tags:
        - Media
      summary: Upload media file
      description: Upload image or video files with processing
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image or video file (max 100MB)
      responses:
        '200':
          description: Media uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUpload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /media:
    get:
      tags:
        - Media
      summary: Get user media
      description: Retrieve media files uploaded by the authenticated user
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: type
          in: query
          schema:
            type: string
            enum: [image, video]
          description: Filter by media type
      responses:
        '200':
          description: Media files retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Reports Endpoints
  /reports:
    get:
      tags:
        - Reports
      summary: Get all reports
      description: Retrieve all user reports (admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved, dismissed]
        - name: type
          in: query
          schema:
            type: string
            enum: [post, media]
      responses:
        '200':
          description: Reports retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Reports
      summary: Create report
      description: Report inappropriate content
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReport'
      responses:
        '201':
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        isAdmin:
          type: boolean
          default: false
    Session:
      type: object
      properties:
        id:
          type: string
        expiresAt:
          type: string
          format: date-time
        token:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        image:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        totalPoints:
          type: integer
          minimum: 0
        role:
          type: string
          enum: [player, parent]
        childId:
          type: string
          nullable: true
        isOnboarded:
          type: boolean
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
          default: false
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProfile:
      type: object
      properties:
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [player, parent]
          default: player
        childId:
          type: string
          format: uuid
        jerseyNumber:
          type: integer
          minimum: 1
          maximum: 99
        isOnboarded:
          type: boolean
          default: false
        totalPoints:
          type: integer
          minimum: 0
          default: 0
...
    Workout:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        exerciseType:
          type: string
          enum: [dribbling, shooting, conditioning]
        pointsEarned:
          type: integer
          minimum: 0
        durationMinutes:
          type: integer
          minimum: 1
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateWorkout:
      type: object
      required:
        - exerciseType
        - pointsEarned
        - durationMinutes
      properties:
        exerciseType:
          type: string
          enum: [dribbling, shooting, conditioning]
        pointsEarned:
          type: integer
          minimum: 0
        durationMinutes:
          type: integer
          minimum: 1
        notes:
          type: string
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        workoutId:
          type: string
          format: uuid
          nullable: true
        gameId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
    PostWithRelations:
      allOf:
        - $ref: '#/components/schemas/Post'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                  nullable: true
                avatarUrl:
                  type: string
                  nullable: true
                jerseyNumber:
                  type: integer
                  nullable: true
            workout:
              $ref: '#/components/schemas/Workout'
            game:
              $ref: '#/components/schemas/Game'
            media:
              type: array
              items:
                $ref: '#/components/schemas/MediaFile'
            likes:
              type: array
              items:
                type: object
                properties:
                  userId:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      avatarUrl:
                        type: string
                        nullable: true
            comments:
              type: array
              items:
                $ref: '#/components/schemas/CommentWithUser'
            likesCount:
              type: integer
            commentsCount:
              type: integer
            userHasLiked:
              type: boolean
    CreatePost:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
        imageUrl:
          type: string
          format: uri
        workoutId:
          type: string
          format: uuid
        gameId:
          type: string
          format: uuid
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        userId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentWithUser:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                  nullable: true
                avatarUrl:
                  type: string
                  nullable: true
...
    FileUpload:
      type: object
      properties:
        url:
          type: string
          format: uri
        name:
          type: string
        size:
          type: integer
        type:
          type: string
    MediaUpload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        fileName:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        processingStatus:
          type: string
          enum: [processing, completed, failed]
    MediaFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        url:
          type: string
          format: uri
        fileName:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        processingStatus:
          type: string
          enum: [processing, completed, failed]
        createdAt:
          type: string
          format: date-time
...
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        isSharedToFeed:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateGame:
      type: object
      required:
        - teamName
        - isHome
        - opponentTeam
        - gameDate
      properties:
        teamName:
          type: string
        isHome:
          type: boolean
        opponentTeam:
          type: string
        gameDate:
          type: string
          format: date-time
        notes:
          type: string
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporterId:
          type: string
        reportType:
          type: string
          enum: [post, media]
        reportedItemId:
          type: string
          format: uuid
        reason:
          type: string
        status:
          type: string
          enum: [pending, resolved, dismissed]
          default: pending
        createdAt:
          type: string
          format: date-time
    CreateReport:
      type: object
      required:
        - reportType
        - reportedItemId
        - reason
      properties:
        reportType:
          type: string
          enum: [post, media]
        reportedItemId:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 1
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid data provided
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Insufficient permissions
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
...